refactor(trade_executor): Move trade execution logic to trade executor
